requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
fastapi==0.115.8
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
idna==3.10
limits==4.0.1
oauthlib==3.2.2
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
